Refactor Periodic into a base class
We'll borrow a lot of the functionality when making the forgetting
version, so we just let the concrete class define it's own UCB method
and pull the rest of the functionality into the base.

For the forgetting agent, we'll need the full history of the conversions
(which is also sufficient for the regular Periodic agent), so we
refactor to allow that.